build: support multiple target systems

+++ de.kennyhml.e4.abap_highlighter/src/main/native/de_kennyhml_e4_abap_highlighter_TreeSitterAbap.c
@@ -1,4 +1,5 @@
 #include <jni.h>
+#include <stdint.h>
 
 void *tree_sitter_abap();
 /*
@@ -8,5 +9,5 @@ void *tree_sitter_abap();
  */
 JNIEXPORT jlong JNICALL Java_de_kennyhml_e4_abap_1highlighter_TreeSitterAbap_tree_1sitter_1abap
   (JNIEnv *env, jclass clz){
-   return (jlong) tree_sitter_abap();
+   return (jlong)(intptr_t)tree_sitter_abap();
 }
